feat: start identity verification without authenticating (#4133)
* startIdentityVerificationMutation does not require auth

* remove identityVerification by id loader with auth

* adjust tests

* add back the guard but return null

* remove unecessary loader check
